refactor: import directly from @threa/types, remove next_message_only
- Remove re-export layer (apps/backend/src/lib/constants.ts)
- Backend now imports constants/types directly from @threa/types
- Create schemas.ts for Zod validation schemas (backend-only concern)
- Remove next_message_only from COMPANION_MODES constant
- Update frontend db types and test client accordingly
